Added setcap target to adjust tools capabilities.
Some unit tests or tools need more capabilities to open a PF_PACKET
socket for instance. It is either possible to run them with higher
privileges (e.g with sudo) or only grant them the needed capabilities
(with setcap) and run them as normal user.

The latter gives more flexibility as you can only grant the needed
capabilities and nothing more.

To do that, a new target called "setcap" which only set capabilities.
It must be run with higher permissions in order to work.

Example:
 $ cd build/
 $ cmake ..
 $ make
 $ sudo make setcap
 $ make test