Break build if setcap is not present.
Not having setcap on a system to grant capabilities is not option.
Created IPC handles with elevated privileges means that they can only be
removed with the same amount of privileges.
It would be possible to get rid of those with ipcrm(1) but it not worth
the pain at all...