Grant capabilites to dabbad on install.
The proper way to do it would be to create a dependency on the
'install'.
However, 'install'/'test'/'all' are built-in targets and are only created
at the very end of the makefile generate.
The direct consequence is that we cannot use ADD_CUSTOM_TARGET() to add
tasks on 'install'.

This workaround tells CMake to append this CMake command directly to
the 'install' target.
As 'make install' must be called with higher privileges to perform
properly, the 'setcap' process uses this context as well.