ipc: mark IPC message origin.
The daemon was first configured to receive the first incoming
message from the IPC queue. However, when the client was too slow
to dequeue a reply from the daemon, the daemon would take again the
same message and try to process it again.
It led to invalid output on the client side.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@gmail.com>