dabbad: add link_destroy wrapper function.
unlike free(3) which does not perform any operation if a NULL pointer is
given. nl_object_free() will try to access the object's operations which
causes a segfault when a NULL pointer is given.
The new wrapper function link_destroy() takes a pointer to struct
rtnl_link and frees it if non-NULL.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@gmail.com>