cmake: do not install libdabba shlib symlink.
Debian packages need to have the shlib version as prefix.
CMake allows to add it and generates a symlink to build the project:

$ stat libdabba.so
  File: `libdabba.so' -> `libdabba.so.0.3'
  [...]

However, the symlink is installed within the deb package and lintian sees
there something is off:
$ lintian libdabba_0.3_amd64.deb
E: libdabba0.3: md5sums-lists-nonexisting-file usr/lib/libdabba.so
W: libdabba0.3: non-dev-pkg-with-shlib-symlink usr/lib/libdabba.so.0.3
usr/lib/libdabba.so

With this change uses the symlink to build the project but does not
add in the install target.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@gmail.com>